split code in diffent files and classes, deleted test.rb
